dodat mail servis i autorizacija

+++ Cultural-content/src/main/java/com/kts/cultural_content/repository/VerificationTokenRepository.java
@@ -4,6 +4,6 @@ import com.kts.cultural_content.model.Admin;
 import com.kts.cultural_content.model.VerificationToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Long> {
+public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Integer> {
     VerificationToken findByToken(String url);
 }
